query fixed

+++ ProyectoFinalTomasDelPino-1/src/main/java/com/salesianostriana/dam/models/Usuario.java
@@ -41,7 +41,7 @@ public class Usuario implements UserDetails{
 	private boolean isAdmin;
 	
 	@Column(name = "num_ofertas", columnDefinition = "SMALLINT DEFAULT 0")
-	private int num_ofertas;
+	private long num_ofertas;
 	
 	@Column(name = "calificacion", columnDefinition = "NUMERIC(3,2) DEFAULT 0")
 	private float calificacion;
